test/WBFdb/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==5.2.5
django-allauth==65.10.0
django-cors-headers==4.7.0
django-extensions==4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.0
drf-access-policy==1.5.0
drf-extra-fields==3.7.0
drf-nested-routers==0.94.2
filetype==1.2.0
greenlet==3.2.3
idna==3.10
iniconfig==2.1.0
isodate==0.7.2
lxml==6.0.0
msgpack==1.1.1
numpy==2.3.2
oauthlib==3.3.1
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.2
PyJWT==2.9.0
pynvim==0.5.2
pyparsing==3.2.3
pytest==8.4.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
python3-saml==1.16.0
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
rest-pandas==1.1.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
xmlsec==1.3.16
